The Chi-bar-squared Distribution

Density, distribution function, and random generation for the chi-bar-squared distribution with mixing weights (conic intrinsic volumes) v.

dchibarsq(x, v)

pchibarsq(q, v)

rchibarsq(n, v)

Arguments

x, q
vector of quantiles.

v
vector of mixing weights (conic intrinsic volumes).

n
number of observations.

Value

The output of dchibarsq(x, v) is a vector of length length(x) such that the entries are the values of the density of the chi-bar-squared distribution with weights v in the corresponding entry of x.

The output of pchibarsq(q, v) is a vector of length length(q) such that the entries are the values of the cdf of the chi-bar-squared distribution with weights v in the corresponding entry of q.

The output of rchibarsq(n, v) is a vector of length n such that the entries are iid samples from the chi-bar-squared distribution with weights v.

Functions

dchibarsq: Evaluates the density of the chi-bar-squared distribution.

pchibarsq: Evaluates the cdf of the chi-bar-squared distribution.

rchibarsq: Samples from the chi-bar-squared distribution.

See also

dbichibarsq, pbichibarsq, rbichibarsq, dchisq, pchisq, qchisq, rchisq

Package: conivol

Examples

dchibarsq(3,c(0.1,0.6,0.3))

#> [1] 0.07207839

dchibarsq(c(0.2,3,4.5),c(0.1,0.6,0.3))

#> [1] 0.35216834 0.07207839 0.03360194

pchibarsq(3,c(0.1,0.6,0.3))

#> [1] 0.6577955

pchibarsq(c(0.2,3,4.5),c(0.1,0.6,0.3))

#> [1] 0.1916255 0.6577955 0.7333710

rchibarsq(10,c(0.1,0.6,0.3))

#>  [1] 0.4573625 1.0273051 1.0422421 0.3612420 2.1892714 0.2805944 0.1731718
#>  [8] 1.6605553 1.6844273 4.1034157