conivol: A package for the (bivariate) chi-bar-squared distribution and conic intrinsic volumes

The conivol package provides functions for the chi-bar-squared distribution, the bivariate chi-bar-squared distribution, and the conic intrinsic volumes.

Details

conivol supports standard functions for the density/cdf/sampling of the (bivariate) chi-bar-squared distribution, calculations and known formulas for special classes of intrinsic volumes of cones, sampling functions for ellipsoidal cones and general polyhedral cones, as well as functions for estimating intrinsic volumes either from direct samples of the intrinsic volumes distribution (in the case of polyhedral cones) or from samples of the corresponding bivariate chi-bar-squared distribution. The package supports point estimates as well as Bayesian estimates via JAGS and Stan.

(Bivariate) Chi-bar-squared distribution

dchibarsq / pchibarsq / rchibarsq: evaluates the density / evaluates the cumulative distribution function / produces samples of the chi-bar-squared distribution

dbichibarsq / pbichibarsq / rbichibarsq: evaluates the density / evaluates the cumulative distribution function / produces samples of the bivariate chi-bar-squared distribution

Special classes of cones

prod_ivols: computes the intrinsic volumes of a product cone by convolving the intrinsic volumes of its elements

circ_ivols: computes the intrinsic volumes of (a product of) circular cones

ellips_semiax / ellips_rbichibarsq: computes the semiaxes / produces samples from the bivariate chi-bar-squared distribution of an ellipsoidal cone

weyl_matrix / weyl_ivols: computes a matrix representation / computes the intrinsic volumes of (a product of) Weyl chambers

General polyhedral cones

polyh_reduce_gen / polyh_reduce_ineq: compute a reduced representation of a polyhedral cone given by generators / inequalities

polyh_rivols_gen / polyh_rivols_ineq: produce samples from the intrinsic volumes distribution of a polyhedral cone given by generators / inequalities

polyh_rbichibarsq_gen / polyh_rbichibarsq_ineq: produce samples from the bivariate chi-bar-squared distribution with weights given by the conic intrinsic volumes of a polyhedral cone given by generators / inequalities

polyh_bayes: generates functions for computing quantiles of marginals of the posterior distribution and for sampling from the posterior distribution, given samples of the intrinsic volumes distribution (based on analytic solution)

polyh_stan: generates inputs for Stan (data list and model string or external file) for sampling from the posterior distribution, given samples of the intrinsic volumes distribution using a model that naturally implies log-concavity (and cannot be solved analytically)

Estimating the weights of the bivariate chi-bar-squared distribution

estim_statdim_var: estimates the statistical dimension and the variance of the intrinsic volumes from samples of the corresponding bivariate chi-bar-squared distribution

init_ivols: finds an initial estimate of the weights, potentially based on first and/or second moment

loglike_ivols: computes the log-likelihood of a weight vector for specific sample data

prepare_em: evaluates the sample data of the bivariate chi-bar-squared data (find the corresponding chi-squared density values)

estim_em: produces EM-type iterates that may or may not converge to the maximum likelihood estimate for the weights of the bivariate chi-bar-squared distribution from sample data

estim_jags / estim_stan: generate inputs for JAGS / Stan (data list and model string or external file) for sampling from the posterior distribution of the intrinsic volumes, given samples of the bivariate chi-bar-squared distribution

See also

github

github.com/damelunx/conivol

vignette

Conic intrinsic volumes and (bivariate) chi-bar-squared distribution: introduces conic intrinsic volumes and (bivariate) chi-bar-squared distributions, as well as the computations involving polyhedral cones

vignette

Estimating conic intrinsic volumes from bivariate chi-bar-squared data: describes the details of the algorithm for finding the intrinsic volumes of closed convex cones from samples of the associated bivariate chi-bar-squared distribution

vignette

Bayesian estimates for conic intrinsic volumes: describes the Bayesian approach for reconstructing intrinsic volumes from sampling data, which can either be samples from the intrinsic volumes distribution (in the case of polyhedral cones), or from the bivariate chi-bar-squared distribution, and which can be with or without enforcing log-concavity of the intrinsic volumes