Computes the semiaxes of an ellipsoidal cone given as the linear image of the Lorentz cone

ellips_semiax takes as input an invertible matrix A and returns the lengths of the semiaxes of the ellipsoidal cone given by the image of the Lorentz cone under A.

ellips_semiax(A)

Arguments

A
matrix

Value

The output of ellips_semiax(A), A in Gl_d, is a positive vector a in R^(d-1) such that the cone A*L, where L={x in R^d | sqrt(2)*x_d >= ||x||}, is isometric to the cone {x in R^d | x_d >= sum_(j=1)^(d-1) x_j^2/a_j^2}.

Note

See this vignette for further info.

See also

ellips_rbichibarsq

Package: conivol

Examples

A <- matrix(c(2,3,5,7,11,13,17,19,23),3,3)
ellips_semiax(A)

#> [1] 0.10984660 0.03065619