Finding an initial estimate of the intrinsic volumes

init_ivols find an initial estimate of the intrinsic volumes via moment-fitting.

init_ivols(d, init_mode = 0, delta = d/2, var = d/4)

Arguments

d
the dimension of the bivariate chi-bar squared distribution.

init_mode
specifies the way through which the initial estimate is found:

init_mode==0: uniform distribution

init_mode==1: discretized normal distribution

init_mode==2: circular cone fitting statdim

init_mode==3: circular cone fitting variance

init_mode==4: circular cone fitting statdim and variance (geometric mean)

delta
an estimate of the statistical dimension of the cone.

var
an estimate of the variane of the intrinsic volumes.

Value

The output of init_ivols is a (d+1)-column vector.

See also

rbichibarsq, circ_rbichibarsq, rbichibarsq_polyh, loglike_ivols

Package: conivol

Examples

D <- c(5,5)
d <- sum(D)
alpha <- c(pi/3,pi/4)
v_exact <- circ_ivols(D, alpha, product=TRUE)
m_samp <- rbichibarsq(10^6,v)

#> Error in conivol:::.test_vector(v): object 'v' not found

est <- estim_statdim_var(d, m_samp)

#> Error in is.data.frame(x): object 'm_samp' not found

list( v_exact = v_exact , v_init0 = init_ivols( d ) ,
      v_init1 = init_ivols( d, 1, delta=est$delta, var=est$var ) ,
      v_init2 = init_ivols( d, 2, delta=est$delta ) ,
      v_init3 = init_ivols( d, 3, delta=est$delta, var=est$var ) ,
      v_init4 = init_ivols( d, 4, delta=est$delta, var=est$var ) )

#> Error in pnorm((k + 0.5 - delta)/sqrt(var)): object 'est' not found