The conic intrinsic volumes of (products/duals of) Weyl chambers

weyl_ivols computes the conic intrinsic volumes of (polars of) Weyl chambers of finite reflection groups of type A, BC, D.

weyl_ivols(d, cone_type, product = FALSE)

Arguments

d
vector of dimensions; must be same length as cone_type.

cone_type
vector of cone types; must be same length as d, the available types are as follows:

"A":

chamber of type A

"A_red":

chamber of type A, reduced form

"BC":

chamber of type BC

"D":

chamber of type D

"Ap":

polar of chamber of type A

"Ap_red":

polar of chamber of type A, reduced form

"BCp":

polar of chamber of type BC

"Dp":

polar of chamber of type D

product
logical; if TRUE, intrinsic volumes of product cone are returned.

Value

If length(d)==1 or (length(d)>1 & product==TRUE) then a single vectors will be returned. If (length(d)>1 & product==FALSE) then a list of vectors will be returned.

Details

The dimensions and types are given in the vectors d and cone_type (vectors must be of same lengths, entries of conetype must be 'A', 'BC', 'D', 'Ap', 'BCp', or 'Dp'). If the length of the vectors is one, a single vector is returned; if the length of the vectors is greater than one and product==FALSE, a list of vectors is returned; if the length of the vectors is greater than one and product==TRUE, a single vector with the intrinsic volumes of the product cone is returned.

See also

weyl_matrix

Package: conivol

Examples

weyl_ivols(5, "BC")

#> [1] 0.2460937500 0.4398437500 0.2473958333 0.0598958333 0.0065104167
#> [6] 0.0002604167

weyl_ivols(c(5,5), c("BC","Ap"))

#> [[1]]
#> [1] 0.2460937500 0.4398437500 0.2473958333 0.0598958333 0.0065104167
#> [6] 0.0002604167
#> 
#> [[2]]
#> [1] 0.001388889 0.020833333 0.118055556 0.312500000 0.380555556 0.166666667
#> [7] 0.000000000
#> 

weyl_ivols(c(5,5), c("BC","Ap"), product = TRUE)

#>  [1] 3.417969e-04 5.737847e-03 3.855975e-02 1.340676e-01 2.615668e-01
#>  [6] 2.929188e-01 1.869466e-01 6.609158e-02 1.254159e-02 1.184172e-03
#> [11] 4.340278e-05 0.000000e+00

# testing the reduced cones
d <- 6
weyl_ivols(d, "A")

#> [1] 0.0000000000 0.1428571429 0.3500000000 0.3222222222 0.1458333333
#> [6] 0.0347222222 0.0041666667 0.0001984127

weyl_ivols(d, "A_red")

#> [1] 0.1428571429 0.3500000000 0.3222222222 0.1458333333 0.0347222222
#> [6] 0.0041666667 0.0001984127

weyl_ivols(d, "Ap")

#> [1] 0.0001984127 0.0041666667 0.0347222222 0.1458333333 0.3222222222
#> [6] 0.3500000000 0.1428571429 0.0000000000

weyl_ivols(d, "Ap_red")

#> [1] 0.0001984127 0.0041666667 0.0347222222 0.1458333333 0.3222222222
#> [6] 0.3500000000 0.1428571429